requirements.txt
Django==2.2.24
psycopg2-binary==2.9.1
Faker==8.11.0
ipdb==0.13.9
django-dotenv==1.4.2
# profilers
nplusone==1.0.0
django-silk==4.1.0
django-postgres-metrics==0.10.1
django-cprofile-middleware==1.0.5
django-debug-toolbar==3.2.2
django-health-check==3.16.4
django-auto-prefetch==0.0.6
# rest
django-filter==2.4.0
django-cors-headers==3.5.0
djangorestframework==3.12.4
django-auto-prefetching==0.1.10 
# graphql
graphene-django==2.15.0